Remove auth id from prod queried fields config.
In da04bea, authoritative_id was added to the configuration for all
environments, but it should have only been added for integration and qa
until the story is accepted. Nothing has been erroneously deployed to
staging, and this ensures that if dataset-search-services is deployed to
staging, authoritative_id will not be in the queried fields before the
story is accepted.

[Trello #1039]
[Pivotal 77640426]